requirements.txt
streamlit>=1.30
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
geoalchemy2>=0.14
geopy>=2.4
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
google-genai>=1.0
